feat:레벨 시스템 부분 기능 추가

+++ TextRPG_13/PlayerStatement.cs
@@ -12,8 +12,8 @@ namespace TextRPG_13
         public string Name { get; set; }
         public JOBTYPE Job { get; set; }
         public int Level { get; set; } = 1;
-        public int Offensivepower { get; set; }
-        public int Defensivepower { get; set; }
+        public float Offensivepower { get; set; }
+        public float Defensivepower { get; set; }
         public int Max_HP { get; set; }
         public int HP { get; set; } = 100;
         public int Gold { get; set; } = 1500;
